Revamped the scene loading and cloning which led to a lot of other changes
We were having trouble getting the notification posts to work.
Revamping the scene loading strategy was key here. We're now
loading a single instance of the scene from the reality composer
file. We then add that scene to the ARView's scene. At this point
we are able to clone the plant sign from that scene which is
loaded and added to the ARView Scene.. Any other combination of
cloning or reloading did not work.

Added the helper UI to guide the user through AR orientation
Revamped some of the UI to be a bit cleaner structure.

Commented out the loading from the network just to keep the ARWorld clean