Extract the ViewModel into the Pointfree composable architecture guidelines
The ViewModel is a ton thinner now and most notably the networking responsibilities
have been moved out. We are still in a bit of a messy stage as the networking
is scattered all around. We need to create a coordinator type thing so
that the network client can update the state in a silo'd area.